Isolating API for Rankup
