+++ build.gradle
@@ -21,6 +21,9 @@ targetCompatibility = '1.8'
 
 repositories {
     mavenCentral()
+    flatDir {
+        dirs "libs"
+    }
     maven {
         name = 'placeholderapi-repo'
         url = 'https://repo.extendedclip.com/content/repositories/placeholderapi/'
@@ -42,12 +45,6 @@ repositories {
         url = "https://repo.codemc.io/repository/nms/"
     }
     maven {
-        name = "worldguard because mcmmo is a needy dependency"
-        url = "https://maven.enginehub.org/repo/"
-    }
-
-    // Added for Factions
-    maven {
         name = "factions repo"
         url = "https://repo.dustplanet.de/artifactory/ext-release-local/"
     }
@@ -59,13 +56,19 @@ repositories {
 
 dependencies {
     implementation "org.spigotmc:spigot:$spigotVersion"
-    implementation "net.milkbowl.vault:VaultAPI:$vaultVersion"
-    implementation "me.clip:placeholderapi:$placeholderVersion"
-    implementation "com.gmail.nossr50.mcMMO:mcMMO:$mcmmoVersion"
+    implementation ("net.milkbowl.vault:VaultAPI:$vaultVersion") {
+        exclude group: "org.bukkit"
+    }
+    implementation ("me.clip:placeholderapi:$placeholderVersion") {
+        exclude group: "org.bstats"
+    }
+    implementation ("com.github.mcMMO-Dev:mcMMO:$mcmmoVersion") {
+        exclude group: "com.sk89q.worldguard"
+    }
     implementation "com.SirBlobman.combatlogx:CombatLogX-API:$combatlogxVersion"
     implementation "com.massivecraft:Factions:$factionsVersion"
     implementation "com.github.brcdev-minecraft:shopgui-api:$shopguiplusVersion"
-    implementation "com.github.okx-code:Rankup3:master" // Pulls from jitpack.io
+    implementation "sh.okx.rankup:Rankup:3.12.2"
 
     testCompile "junit:junit:4.12"
 }
